Implementing the `run-deployment` automation action
This brings the `run-deployment` action to parity with Prefect Cloud, which
required bringing quite a few additional things along.

First, we needed to bring the Jinja2 base `Action` and refresh the open-source
version of schema hydration to include Jinja2 support.  This will aid with
several upcoming `Action`s as well.

Second, we needed the support for `created_by`, which was stubbed in for OSS,
but had a fuller implementation in Prefect Cloud.  I've ported the automations
part of that to OSS.

I'm also removing a long-standing TODO to allow some configuration of whether
our custom client will always raise on HTTP errors or not.  The actions template
system expected to get raw responses (and test for 404s, etc).  Making this an
argument to the Prefect HTTP client allows us to change this behavior.